refactor charts and RegionSelectionWrapper
